+++ databricks-sdk-java/src/main/java/com/databricks/sdk/service/catalog/ListSystemSchemasResponse.java
@@ -10,9 +10,25 @@ import java.util.Objects;
 
 @Generated
 public class ListSystemSchemasResponse {
+  /**
+   * Opaque token to retrieve the next page of results. Absent if there are no more pages.
+   * __page_token__ should be set to this value for the next request (for the next page of results).
+   */
+  @JsonProperty("next_page_token")
+  private String nextPageToken;
+
   /** An array of system schema information objects. */
   @JsonProperty("schemas")
   private Collection<SystemSchemaInfo> schemas;
+
+  public ListSystemSchemasResponse setNextPageToken(String nextPageToken) {
+    this.nextPageToken = nextPageToken;
+    return this;
+  }
+
+  public String getNextPageToken() {
+    return nextPageToken;
+  }
 
   public ListSystemSchemasResponse setSchemas(Collection<SystemSchemaInfo> schemas) {
     this.schemas = schemas;
@@ -28,16 +44,20 @@ public class ListSystemSchemasResponse {
     if (this == o) return true;
     if (o == null || getClass() != o.getClass()) return false;
     ListSystemSchemasResponse that = (ListSystemSchemasResponse) o;
-    return Objects.equals(schemas, that.schemas);
+    return Objects.equals(nextPageToken, that.nextPageToken)
+        && Objects.equals(schemas, that.schemas);
   }
 
   @Override
   public int hashCode() {
-    return Objects.hash(schemas);
+    return Objects.hash(nextPageToken, schemas);
   }
 
   @Override
   public String toString() {
-    return new ToStringer(ListSystemSchemasResponse.class).add("schemas", schemas).toString();
+    return new ToStringer(ListSystemSchemasResponse.class)
+        .add("nextPageToken", nextPageToken)
+        .add("schemas", schemas)
+        .toString();
   }
 }