+++ databricks-sdk-java/src/main/java/com/databricks/sdk/service/sharing/UpdateRecipient.java
@@ -13,6 +13,10 @@ public class UpdateRecipient {
   @JsonProperty("comment")
   private String comment;
 
+  /** Expiration timestamp of the token, in epoch milliseconds. */
+  @JsonProperty("expiration_time")
+  private Long expirationTime;
+
   /** IP Access List */
   @JsonProperty("ip_access_list")
   private IpAccessList ipAccessList;
@@ -43,6 +47,15 @@ public class UpdateRecipient {
 
   public String getComment() {
     return comment;
+  }
+
+  public UpdateRecipient setExpirationTime(Long expirationTime) {
+    this.expirationTime = expirationTime;
+    return this;
+  }
+
+  public Long getExpirationTime() {
+    return expirationTime;
   }
 
   public UpdateRecipient setIpAccessList(IpAccessList ipAccessList) {
@@ -96,6 +109,7 @@ public class UpdateRecipient {
     if (o == null || getClass() != o.getClass()) return false;
     UpdateRecipient that = (UpdateRecipient) o;
     return Objects.equals(comment, that.comment)
+        && Objects.equals(expirationTime, that.expirationTime)
         && Objects.equals(ipAccessList, that.ipAccessList)
         && Objects.equals(name, that.name)
         && Objects.equals(newName, that.newName)
@@ -105,13 +119,15 @@ public class UpdateRecipient {
 
   @Override
   public int hashCode() {
-    return Objects.hash(comment, ipAccessList, name, newName, owner, propertiesKvpairs);
+    return Objects.hash(
+        comment, expirationTime, ipAccessList, name, newName, owner, propertiesKvpairs);
   }
 
   @Override
   public String toString() {
     return new ToStringer(UpdateRecipient.class)
         .add("comment", comment)
+        .add("expirationTime", expirationTime)
         .add("ipAccessList", ipAccessList)
         .add("name", name)
         .add("newName", newName)